+++ sdk/resourcemanager/advisor/armadvisor/zz_generated_pagers.go
@@ -16,60 +16,6 @@ import (
 	"reflect"
 )
 
-// ConfigurationsClientListBySubscriptionPager provides operations for iterating over paged responses.
-type ConfigurationsClientListBySubscriptionPager struct {
-	client    *ConfigurationsClient
-	current   ConfigurationsClientListBySubscriptionResponse
-	err       error
-	requester func(context.Context) (*policy.Request, error)
-	advancer  func(context.Context, ConfigurationsClientListBySubscriptionResponse) (*policy.Request, error)
-}
-
-// Err returns the last error encountered while paging.
-func (p *ConfigurationsClientListBySubscriptionPager) Err() error {
-	return p.err
-}
-
-// NextPage returns true if the pager advanced to the next page.
-// Returns false if there are no more pages or an error occurred.
-func (p *ConfigurationsClientListBySubscriptionPager) NextPage(ctx context.Context) bool {
-	var req *policy.Request
-	var err error
-	if !reflect.ValueOf(p.current).IsZero() {
-		if p.current.ConfigurationListResult.NextLink == nil || len(*p.current.ConfigurationListResult.NextLink) == 0 {
-			return false
-		}
-		req, err = p.advancer(ctx, p.current)
-	} else {
-		req, err = p.requester(ctx)
-	}
-	if err != nil {
-		p.err = err
-		return false
-	}
-	resp, err := p.client.pl.Do(req)
-	if err != nil {
-		p.err = err
-		return false
-	}
-	if !runtime.HasStatusCode(resp, http.StatusOK) {
-		p.err = runtime.NewResponseError(resp)
-		return false
-	}
-	result, err := p.client.listBySubscriptionHandleResponse(resp)
-	if err != nil {
-		p.err = err
-		return false
-	}
-	p.current = result
-	return true
-}
-
-// PageResponse returns the current ConfigurationsClientListBySubscriptionResponse page.
-func (p *ConfigurationsClientListBySubscriptionPager) PageResponse() ConfigurationsClientListBySubscriptionResponse {
-	return p.current
-}
-
 // OperationsClientListPager provides operations for iterating over paged responses.
 type OperationsClientListPager struct {
 	client    *OperationsClient
@@ -124,60 +70,6 @@ func (p *OperationsClientListPager) PageResponse() OperationsClientListResponse 
 	return p.current
 }
 
-// RecommendationMetadataClientListPager provides operations for iterating over paged responses.
-type RecommendationMetadataClientListPager struct {
-	client    *RecommendationMetadataClient
-	current   RecommendationMetadataClientListResponse
-	err       error
-	requester func(context.Context) (*policy.Request, error)
-	advancer  func(context.Context, RecommendationMetadataClientListResponse) (*policy.Request, error)
-}
-
-// Err returns the last error encountered while paging.
-func (p *RecommendationMetadataClientListPager) Err() error {
-	return p.err
-}
-
-// NextPage returns true if the pager advanced to the next page.
-// Returns false if there are no more pages or an error occurred.
-func (p *RecommendationMetadataClientListPager) NextPage(ctx context.Context) bool {
-	var req *policy.Request
-	var err error
-	if !reflect.ValueOf(p.current).IsZero() {
-		if p.current.MetadataEntityListResult.NextLink == nil || len(*p.current.MetadataEntityListResult.NextLink) == 0 {
-			return false
-		}
-		req, err = p.advancer(ctx, p.current)
-	} else {
-		req, err = p.requester(ctx)
-	}
-	if err != nil {
-		p.err = err
-		return false
-	}
-	resp, err := p.client.pl.Do(req)
-	if err != nil {
-		p.err = err
-		return false
-	}
-	if !runtime.HasStatusCode(resp, http.StatusOK) {
-		p.err = runtime.NewResponseError(resp)
-		return false
-	}
-	result, err := p.client.listHandleResponse(resp)
-	if err != nil {
-		p.err = err
-		return false
-	}
-	p.current = result
-	return true
-}
-
-// PageResponse returns the current RecommendationMetadataClientListResponse page.
-func (p *RecommendationMetadataClientListPager) PageResponse() RecommendationMetadataClientListResponse {
-	return p.current
-}
-
 // RecommendationsClientListPager provides operations for iterating over paged responses.
 type RecommendationsClientListPager struct {
 	client    *RecommendationsClient
@@ -229,59 +121,5 @@ func (p *RecommendationsClientListPager) NextPage(ctx context.Context) bool {
 
 // PageResponse returns the current RecommendationsClientListResponse page.
 func (p *RecommendationsClientListPager) PageResponse() RecommendationsClientListResponse {
-	return p.current
-}
-
-// SuppressionsClientListPager provides operations for iterating over paged responses.
-type SuppressionsClientListPager struct {
-	client    *SuppressionsClient
-	current   SuppressionsClientListResponse
-	err       error
-	requester func(context.Context) (*policy.Request, error)
-	advancer  func(context.Context, SuppressionsClientListResponse) (*policy.Request, error)
-}
-
-// Err returns the last error encountered while paging.
-func (p *SuppressionsClientListPager) Err() error {
-	return p.err
-}
-
-// NextPage returns true if the pager advanced to the next page.
-// Returns false if there are no more pages or an error occurred.
-func (p *SuppressionsClientListPager) NextPage(ctx context.Context) bool {
-	var req *policy.Request
-	var err error
-	if !reflect.ValueOf(p.current).IsZero() {
-		if p.current.SuppressionContractListResult.NextLink == nil || len(*p.current.SuppressionContractListResult.NextLink) == 0 {
-			return false
-		}
-		req, err = p.advancer(ctx, p.current)
-	} else {
-		req, err = p.requester(ctx)
-	}
-	if err != nil {
-		p.err = err
-		return false
-	}
-	resp, err := p.client.pl.Do(req)
-	if err != nil {
-		p.err = err
-		return false
-	}
-	if !runtime.HasStatusCode(resp, http.StatusOK) {
-		p.err = runtime.NewResponseError(resp)
-		return false
-	}
-	result, err := p.client.listHandleResponse(resp)
-	if err != nil {
-		p.err = err
-		return false
-	}
-	p.current = result
-	return true
-}
-
-// PageResponse returns the current SuppressionsClientListResponse page.
-func (p *SuppressionsClientListPager) PageResponse() SuppressionsClientListResponse {
 	return p.current
 }